+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/ResolveChildDefinitionsPassTest.php
@@ -312,28 +312,6 @@ class ResolveChildDefinitionsPassTest extends TestCase
         $this->assertTrue($container->getDefinition('decorated_deprecated_parent')->isDeprecated());
     }
 
-    /**
-     * @group legacy
-     */
-    public function testDecoratedServiceCanOverwriteDeprecatedParentStatus()
-    {
-        $this->expectDeprecation('Since symfony/dependency-injection 5.1: The signature of method "Symfony\Component\DependencyInjection\Definition::setDeprecated()" requires 3 arguments: "string $package, string $version, string $message", not defining them is deprecated.');
-        $this->expectDeprecation('Since symfony/dependency-injection 5.1: Passing a null message to un-deprecate a node is deprecated.');
-
-        $container = new ContainerBuilder();
-        $container->register('deprecated_parent')
-            ->setDeprecated(true)
-        ;
-
-        $container->setDefinition('decorated_deprecated_parent', new ChildDefinition('deprecated_parent'))
-            ->setDeprecated(false)
-        ;
-
-        $this->process($container);
-
-        $this->assertFalse($container->getDefinition('decorated_deprecated_parent')->isDeprecated());
-    }
-
     public function testProcessResolvesAliases()
     {
         $container = new ContainerBuilder();