+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services1-1.php
@@ -34,12 +34,4 @@ class Container extends \Symfony\Component\DependencyInjection\Dump\AbstractCont
     {
         return true;
     }
-
-    public function getRemovedIds(): array
-    {
-        return [
-            'Psr\\Container\\ContainerInterface' => true,
-            'Symfony\\Component\\DependencyInjection\\ContainerInterface' => true,
-        ];
-    }
 }