+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services19.php
@@ -39,14 +39,6 @@ class ProjectServiceContainer extends Container
         return true;
     }
 
-    public function getRemovedIds(): array
-    {
-        return [
-            'Psr\\Container\\ContainerInterface' => true,
-            'Symfony\\Component\\DependencyInjection\\ContainerInterface' => true,
-        ];
-    }
-
     /**
      * Gets the public 'service_from_anonymous_factory' shared service.
      *