+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services33.php
@@ -37,14 +37,6 @@ class ProjectServiceContainer extends Container
         return true;
     }
 
-    public function getRemovedIds(): array
-    {
-        return [
-            'Psr\\Container\\ContainerInterface' => true,
-            'Symfony\\Component\\DependencyInjection\\ContainerInterface' => true,
-        ];
-    }
-
     /**
      * Gets the public 'Bar\Foo' shared service.
      *