+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services8.php
@@ -35,14 +35,6 @@ class ProjectServiceContainer extends Container
         return true;
     }
 
-    public function getRemovedIds(): array
-    {
-        return [
-            'Psr\\Container\\ContainerInterface' => true,
-            'Symfony\\Component\\DependencyInjection\\ContainerInterface' => true,
-        ];
-    }
-
     /**
      * @return array|bool|float|int|string|null
      */